Jb fix vocabulary (#214)
* tweaked the vocabulary logic

* fixed languages
