+++ public/js/contribute/mobilization/main.js
@@ -1,4 +1,4 @@
-import { vocabulary } from '/js/config/translations.js';
+import { vocabulary } from '/js/config/main.js';
 import { fixLabel } from '/js/main.js';
 
 export function adjustarea(node) {
