+++ public/js/contribute/pad/exploration.js
@@ -1,7 +1,7 @@
 import { Exploration } from '/js/exploration.js';
 // INIT EXPLORATION
 
-export function initExploration() {
+function initExploration() {
   const pad = JSON.parse(d3.select('data[name="pad"]')?.node()?.value);
   const exploration = new Exploration();
   exploration.setUseFullPromptForSelect(true);
@@ -15,3 +15,5 @@ export function initExploration() {
     }
   });
 }
+
+export const exploration = initExploration();
