+++ public/js/contribute/pad/load.js
@@ -1,3 +1,4 @@
+import { initGTranslate } from '/js/config/gtranslate.js'
 import { selectReviewLanguage } from '/js/contribute/pad/main.js';
 import { renderPad } from '/js/contribute/pad/render.js';
 import { partialSave, saveAndSubmit } from '/js/contribute/pad/save.js';
@@ -6,6 +7,9 @@ import { getMediaSize } from '/js/main.js';
 
 async function DOMLoad() {
   if (!mediaSize) var mediaSize = getMediaSize();
+  
+  await initGTranslate();
+
   const { id, type, source } = JSON.parse(
     d3.select('data[name="pad"]').node()?.value,
   );
