+++ public/js/contribute/template/keyboard.interactions.js
@@ -1,4 +1,4 @@
-import { language } from '/js/config/translations.js';
+import { language } from '/js/config/main.js';
 import { autofillTitle } from '/js/contribute/template/render.js';
 import { switchButtons } from '/js/contribute/template/save.js';
