+++ public/js/exploration.js
@@ -1,4 +1,4 @@
-import { language, vocabulary } from '/js/config/translations.js';
+import { language, vocabulary } from '/js/config/main.js';
 import { GET, POST, PUT } from '/js/fetch.js';
 import { renderPromiseModal } from '/js/modals.js';
