+++ views/partials/head.ejs
@@ -93,7 +93,7 @@
 <!-- <script type='module' src='/js/modals.js'></script> -->
 <script type='module' src='/js/load.js'></script>
 <script type='module' src='/js/navigate/load.js'></script>
-<script type='module' src='/js/navigate/gtranslate.js'></script>
+<!-- <script type='module' src='/js/navigate/gtranslate.js'></script> -->
 <!-- <script type='text/javascript' src='/js/upload.js'></script> -->
 
 <!-- Google Translate CDN -->
